🔨 refactor(app): Move repeated CastComponent to dedicated component
