+++ app/(tabs)/channel.tsx
@@ -1,145 +1,57 @@
-import { FlashList } from '@shopify/flash-list';
-import { formatDistanceToNow } from 'date-fns';
-import _ from 'lodash';
-import React, { useCallback } from 'react';
-import { View, StyleSheet, ActivityIndicator, Text, Image } from 'react-native';
-import { FontAwesome } from '@expo/vector-icons';
-// import useLatestCasts from '../hooks/useLatestCasts';
-import ComposeCast from '../../components/ComposeCast';
-import { Cast } from '../../providers/NeynarProvider';
-import { Link } from 'expo-router';
-import { useRoute } from '@react-navigation/native';
-import { useLatestCasts } from 'farcasterkit-react-native';
-
-
-const CastComponent = ({ cast }: { cast: Cast }) => {
-  const relativeTime = formatDistanceToNow(new Date(cast.timestamp), { addSuffix: true });
-  return (
-    <View style={styles.castContainer}>
-      <Image source={{ uri: cast.author.pfp_url }} style={styles.pfpImage} />
-      <View style={styles.contentContainer}>
-        <View style={styles.headerContainer}>
-          <Text style={styles.displayName}>{cast.author.display_name}</Text>
-          <Text style={styles.timestamp}>{_.replace(relativeTime, 'about ', '')}</Text>
-        </View>
-        <Text style={styles.castText}>{cast.text}</Text>
-        <View style={styles.reactionsContainer}>
-          <View style={styles.reactionsGroupContainer}>
-            <FontAwesome name="heart-o" size={11} color="black" />
-            <Text style={styles.reactionTextFirst}> {cast.reactions.likes.length}</Text>
-          </View>
-          <View style={styles.reactionsGroupContainer}>
-            <FontAwesome name="retweet" size={11} color="black" />
-            <Text style={styles.reactionText}> {cast.reactions.recasts.length}</Text>
-          </View>
-          <View style={styles.reactionsGroupContainer}>
-            <FontAwesome name="comment-o" size={11} color="black" />
-            <Text style={styles.reactionText}> {cast.replies.count}</Text>
-          </View>
-        </View>
-      </View>
-    </View>
-  );
-};
+import { FlashList } from '@shopify/flash-list'
+import _ from 'lodash'
+import React, { useCallback } from 'react'
+import { View, StyleSheet, ActivityIndicator } from 'react-native'
+import { FontAwesome } from '@expo/vector-icons'
+import ComposeCast from '../../components/ComposeCast'
+import Cast from '../../components/Cast'
+import { useRoute } from '@react-navigation/native'
+import { useLatestCasts } from 'farcasterkit-react-native'
 
 const ChannelScreen = () => {
-  const route = useRoute();
+  const route = useRoute()
   // todo: need to add types/interfaces all of this
-  const { type, parent_url } = route.params;
-  const { casts, isLoading, loadMore, isReachingEnd } = parent_url && parent_url.length > 0 ? useLatestCasts(type as string, parent_url as string) : useLatestCasts(type as string);
+  const { type, parent_url } = route.params
+  const { casts, isLoading, loadMore, isReachingEnd } =
+    parent_url && parent_url.length > 0
+      ? useLatestCasts(type as string, parent_url as string)
+      : useLatestCasts(type as string)
 
   const onEndReached = useCallback(() => {
     if (!isReachingEnd) {
-      loadMore();
+      loadMore()
     }
-  }, [isReachingEnd, loadMore]);
+  }, [isReachingEnd, loadMore])
 
   return (
     <View style={styles.container}>
       <FlashList
         contentContainerStyle={styles.flashList}
         data={casts}
-        renderItem={({ item }) => <Link key={item.hash} href={`/conversation?hash=${item.hash}`}><CastComponent cast={item} /></Link>}
+        renderItem={({ item }) => <Cast key={item.hash} cast={item} />}
         keyExtractor={(item) => item.hash}
         onEndReached={onEndReached}
         onEndReachedThreshold={0.1}
-        ListFooterComponent={() => isLoading && !isReachingEnd ? <ActivityIndicator size="large" color="#000000" /> : null}
+        ListFooterComponent={() =>
+          isLoading && !isReachingEnd ? (
+            <ActivityIndicator size="large" color="#000000" />
+          ) : null
+        }
       />
       <ComposeCast />
     </View>
-  );
-};
+  )
+}
 
 const styles = StyleSheet.create({
-  castContainer: {
-    borderBottomWidth: 1,
-    borderColor: '#eaeaea',
-    flexDirection: 'row',
-    padding: 10,
-    width: '100%',
-  },
-  castText: {
-    color: '#333',
-    fontSize: 14,
-    lineHeight: 18,
-  },
   container: {
     backgroundColor: '#fff',
     flex: 1,
     justifyContent: 'space-between',
   },
-  contentContainer: {
-    flex: 1,
-    justifyContent: 'center',
-  },
-  displayName: {
-    fontWeight: 'bold',
-    marginBottom: 4,
-  },
   flashList: {
     backgroundColor: '#fff',
   },
-  headerContainer: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    width: '100%',
-  },
-  icon: {
-    resizeMode: 'contain',
-  },
-  pfpImage: {
-    borderRadius: 25,
-    height: 30,
-    marginLeft: 10,
-    marginRight: 15,
-    width: 30,
-  },
-  reactionText: {
-    color: '#000',
-    fontSize: 11,
-  },
-  reactionTextFirst: {
-    color: '#000',
-    fontSize: 11,
-  },
-  reactionsContainer: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    justifyContent: 'flex-start',
-    marginTop: 4,
-  },
-  reactionsGroupContainer: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    marginRight: 4,
-  },
-  timestamp: {
-    color: '#666',
-    fontSize: 12,
-    paddingBottom: 4,
-    paddingRight: 6,
-  },
-});
+})
 
-export default ChannelScreen;
+export default ChannelScreen
